Remove any code for multiple dependency support
Even though a package manage that only supports 1 dependency is kind of useless, I see why there's a restriction in place that only 1 is possible in the input. The more I thought about handling multiple dependencies per package, the more it seemed like a rabbit hole that would take quite a bit more time than I'm ready to invest right now.

The solution for a multi-dependency input seems like it would be similar to the single-dependency solution. But instead of recursively building a chain, it would loop over dependencies and recursively build a tree. A tree data structure would be needed with both a way to merge multiple trees/branches together properly as well as flatten the combined tree correctly into an ordered list.